Commit 9
This commit has a lot. Fully completed the player finder so it  can create a list of the players' names to search for later on the sites, making that automatic. Also I've completed all of the bartorvik site, including finding, editing, and screenshotting the chart for the chosen team. I also added a dictionary of all currently known teams that UTC plays against to the program so I could get data on any of those teams now just by changing the chosen team input.