Commit 10
This commit is pretty big. There were a lot of bug fixes and lag proofing done to the main program and I've gotten a few pages of the report formatted with reportlabs. I've kind of left the PyQT work on the backburner and don't know if I'm going to end up doing that unless it becomes necessary because it comes with a much larger learning curve than other things have. Could be a long term project to learn it but I can't really do it in a timely manner and I don't have a real reason to justify putting in that time currently.