Commit 11
finished the formatting for the PDF. about to start working on a way to get the correct images to the right pages. the current method of storing the images was easier at the time to setup but is not going to work for this so I will be changing that next.